+++ CHANGELOG.md
@@ -87,6 +87,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) 
 - Asset regex validity checks are now performed on all platforms instead of just Windows
 - Removed invalid paragraphs from README which still described asset name regexes as suffixes
 - Improved wording in README wording and fixed typos
+- Typos and wrong asset names in known tools
 
 ### Removed